% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{erp_arReceivable_orderno_update}
\alias{erp_arReceivable_orderno_update}
\title{更新销售订单号}
\usage{
erp_arReceivable_orderno_update(erp_token)
}
\arguments{
\item{erp_token}{}
}
\value{
无返回值
}
\description{
更新销售订单号
}
\examples{
erp_arReceivable_orderno_update()
}
